clear/backend/requirements.txt
Flask
flask-cors
google-cloud-speech
pydub
